Tests for expressions.
